simplifies runners imports, moves runtime tests
