+++ dlt/common/runners/__init__.py
@@ -1,4 +1,6 @@
 from . import pool_runner
-from .pool_runner import TRunMetrics, run_pool
+from .pool_runner import run_pool
 from .runnable import Runnable, workermethod
 from .init import initialize_runner
+from .typing import TRunMetrics
+from .venv import Venv, VenvNotFound
