+++ tests/common/runners/test_pipes.py
@@ -4,10 +4,10 @@ from typing import Any, Iterator, NamedTuple
 import pytest
 from dlt.common.exceptions import UnsupportedProcessStartMethodException
 
+from dlt.common.runners import TRunMetrics, Venv
 from dlt.common.runners.stdout import iter_stdout, iter_stdout_with_result
 from dlt.common.runners.synth_pickle import encode_obj, decode_obj, decode_last_obj
-from dlt.common.runners.venv import Venv
-from dlt.common.runtime.telemetry import TRunMetrics
+
 from dlt.common.utils import digest128b
