+++ tests/common/scripts/stdout_encode_exception.py
@@ -1,8 +1,9 @@
 from functools import partial
 from dlt.common.exceptions import UnsupportedProcessStartMethodException
 
+from dlt.common.runners import TRunMetrics
 from dlt.common.runners.stdout import exec_to_stdout
-from dlt.common.runtime.telemetry import TRunMetrics
+
 
 
 def worker(data1, data2):
